// src/api/bindings/Actor.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { Permission } from "./Permission";
import type { Role } from "./Role";

/**
 * The person performing an action.
 */
export type Actor = { userId: string, name: string, role: Role, permissions: Array<Permission>, };