// src/api/bindings/AppStatus.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { UserSummary } from "./UserSummary";

export type AppStatus = { setupComplete: boolean, requireLogin: boolean, autoLockMinutes: number, gymName: string, gymLogo: string | null, users: Array<UserSummary>, clockWarning: string | null, schemaVersion: number, memberCount: number, };